pyproject.toml
[project]
name = "reversible-image-modification-system"
version = "0.1.0"
description = "A distributed system for applying and verifying reversible image modifications"

readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "greenlet>=3.0.0",
    "alembic>=1.16.0",
    "pillow>=11.2.0",
    "numpy>=2.3.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.0",
    "loguru>=0.7.0",
    "tortoise-orm>=0.25.1",
    "aiofiles>=24.1.0",
    "image-modification-algorithms",
    "requests>=2.32.4",
    "requests-toolbelt>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.0",
    "pytest-mock>=3.14.0",
    # Development Tools
    "watchfiles>=1.1.0",
    "poethepoet>=0.35.0",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
]
lint = [
    "ruff>=0.12.0",
    "mypy>=1.16.0",
]
test = [
    "pytest-timeout>=2.4.0",
]

[tool.uv]
default-groups = "all"

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
image-modification-algorithms = { workspace = true }

[tool.ruff]
lint.select = [
    "UP006",  # PEP 585 (list instead of List)
    "UP007",  # PEP 604 (| instead of Union)
    "UP035",  # Deprecated imports
    "F401",   # Unused imports
    "I001",   # Import sorting
]
target-version = "py312"

# Poe Tasks
[tool.poe.tasks]
# Linting and formatting tasks
lint = "uv run ruff check"
lint-fix = "uv run ruff check --fix"
format = "uv run ruff format"
format-check = "uv run ruff format --check"

# Combined tasks
check-all = ["lint", "format-check"]
fix-all = ["lint-fix", "format"]

# Docker tasks
docker-build = "docker compose build"
docker-up = "docker compose up -d"
docker-up-dev = "docker compose -f docker-compose.yml -f docker-compose.dev.yml up"
docker-watch = "docker compose watch"
docker-down = "docker compose down"
docker-logs = "docker compose logs -f"
docker-restart = "docker compose restart"
docker-clean = "docker compose down -v --remove-orphans"

# Test tasks
test = "uv run pytest --ignore=tests/system"
test-verbose = "uv run pytest -v --ignore=tests/system"
test-system = "uv run pytest tests/system/ -v --tb=short"
test-system-fast = "uv run pytest tests/system/ -v --tb=short -m 'not slow'"
test-system-smoke = "uv run pytest tests/system/ -v --tb=short -k 'tiny_image or small_image'"
test-all = "uv run pytest -v"

# Tool Configuration
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "packages/*/tests"]
markers = [
    "slow: marks tests as slow (system tests)",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
